Remove legacy styles
