+++ src/lang/en.rs
@@ -1,5 +1,3 @@
-use crate::types::Numeric;
-
 pub const ARTICLES: [&str; 8] = ["A", "AN", "An", "THE", "The", "a", "an", "the"];
 
 pub const NEVER_CAPITALIZE: [&str; 46] = [
@@ -479,68 +477,3 @@ pub const ALWAYS_CAPITALIZE: [&str; 426] = [
     "zambia",
     "zimbabwe",
 ];
-
-/// Will get an english name for a zero-indexed month.
-/// Returns an error if month out of range.
-pub fn get_month_name(month: u8) -> Result<String, ()> {
-    Ok(match month {
-        0 => "January",
-        1 => "February",
-        2 => "March",
-        3 => "April",
-        4 => "May",
-        5 => "June",
-        6 => "July",
-        7 => "August",
-        8 => "September",
-        9 => "October",
-        10 => "November",
-        11 => "December",
-        _ => return Err(()),
-    }
-    .to_string())
-}
-
-/// Will get an english three-letter abbreviation for a zero-indexed month.
-/// Returns an error if month out of range.
-pub fn get_month_abbr(month: u8, dot: bool) -> Result<String, ()> {
-    let mut res = match month {
-        0 => "Jan",
-        1 => "Feb",
-        2 => "Mar",
-        3 => "Apr",
-        4 => "May",
-        5 => "Jun",
-        6 => "Jul",
-        7 => "Aug",
-        8 => "Sep",
-        9 => "Oct",
-        10 => "Nov",
-        11 => "Dec",
-        _ => return Err(()),
-    }
-    .to_string();
-
-    if dot && month != 4 {
-        res.push('.');
-    }
-
-    Ok(res)
-}
-
-pub fn get_ordinal(num: &Numeric) -> String {
-    match num.single_number() {
-        Some(num) => {
-            let my = num.abs();
-            let suffix = match my {
-                1 => "st",
-                2 => "nd",
-                3 => "rd",
-                _ => "th",
-            };
-
-            num.to_string() + suffix
-        }
-        None => num.to_string(),
-    }
-}
